+++ src/event-dashboards/nodes/register.js
@@ -3,6 +3,7 @@ import { CommandInterpreterNode } from '../../runtime/command-interpreter-node';
 import { JobstatsViewNode } from './jobstats-view-node';
 import { PartitionViewerViewNode } from './partition-viewer-view-node';
 import { LogViewerViewNode } from './logviewer-view-node';
+import { SettingsAuditViewNode } from './settings-audit-view-node';
 import { TopicProbeViewNode } from './topic-probe-view-node';
 import { TopologyManagerViewNode } from './topology-manager-view-node';
 import { WorkerStatusTransformNode } from './worker-status-transform-node';
@@ -12,6 +13,7 @@ CommandInterpreterNode.registerNodeClasses( {
 	JobstatsView: JobstatsViewNode,
 	PartitionViewerView: PartitionViewerViewNode,
 	LogViewerView: LogViewerViewNode,
+	SettingsAuditView: SettingsAuditViewNode,
 	TopicProbeView: TopicProbeViewNode,
 	TopologyManagerView: TopologyManagerViewNode,
 	WorkerStatusTransform: WorkerStatusTransformNode,